fixup! adds a watchdog class implementing canServeSQLRequest()
